// Frontend/src/app/api/models/job-offer-list-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface JobOfferListDto {
  applicationCount?: number;
  dateLimite?: string;
  datePublication?: string;
  id: number;
  jobStatus?: 'DRAFT' | 'PUBLISHED' | 'SUSPENDED' | 'ARCHIVED';
  jobTitle?: string;
  jobType?: 'PER' | 'PATS' | 'CONTRACTUEL';
  typeContrat?: 'CDD' | 'CDI' | 'STAGE';
}